+++ .eslintrc.js
@@ -4,5 +4,9 @@ module.exports = {
   rules: {
     '@typescript-eslint/no-empty-function': 'off',
     'no-useless-escape': 'off',
+    'no-cond-assign': 'off'
+  },
+  env: {
+    jest: true
   }
 }