+++ src/frameworks/index.ts
@@ -11,6 +11,7 @@ import NgxTranslateFramework from './ngx-translate'
 import I18nTagFramework from './i18n-tag'
 import VueSFCFramework from './vue-sfc'
 import FlutterFramework from './flutter'
+import EmberFramework from './ember'
 import NextI18nFramework from './next-i18next'
 
 export type PackageDependencies = Partial<Record<PackageFileType, string[]>>
@@ -21,6 +22,7 @@ export const frameworks: Framework[] = [
   new NgxTranslateFramework(),
   new VSCodeFramework(),
   new FlutterFramework(),
+  new EmberFramework(),
   new NextI18nFramework(),
   new I18nextFramework(),
   new I18nTagFramework(),